I think this is close enough to the original Sand Stroke
After spending a lot of time with the original Processing sketch, I think I've got
something that is recreating the effect. I'll probably switch to doing some studies of my
own with this technique now.

I also updated the renderer to using a coroutine so that you can watch it being drawn. It
still spins the fan up, but the browser is more responsive this way as well.